idrac_tsdb/requirements.txt
autopep8==1.5.7
psycopg2==2.7.7
psycopg2-binary==2.9.1
pycodestyle==2.7.0
python-dotenv==0.18.0
toml==0.10.2
